refactor: move fuel consumer and consumer system to domain
