+++ src/libecalc/domain/infrastructure/energy_components/consumer_system/consumer_system.py
